VAreaPaintTargetArea: Don't cast Team to uint when creating control change event

+++ BlitzSniffer/Tracker/Versus/VArea/VAreaPaintTargetArea.cs
@@ -59,7 +59,7 @@ namespace BlitzSniffer.Tracker.Versus.VArea
                 EventTracker.Instance.AddEvent(new VAreaPaintAreaControlChangeEvent()
                 {
                     AreaIdx = AreaIdx,
-                    Team = (uint)team
+                    Team = team
                 });
             }
